+++ components/src/main/java/org/patternfly/component/drawer/DrawerCloseButton.java
@@ -45,7 +45,7 @@ public class DrawerCloseButton extends DrawerSubComponent<HTMLDivElement, Drawer
     DrawerCloseButton() {
         super(SUB_COMPONENT_ID, SUB_COMPONENT_NAME, div().css(component(drawer, actions)).element());
         add(div().css(component(drawer, close))
-                .add(button().plain().icon(IconSets.rhUi.close())
+                .add(button().plain().icon(IconSets.rhMicrons.close())
                         .aria(label, "Close drawer panel")
                         .on(click, e -> close())));
     }